Support for http and https calls
