+++ Cargo.toml
@@ -20,10 +20,13 @@ minimal = []
 
 [dependencies]
 log = { version = "^0.4.22", features = ["std", "kv"] }
-chrono = { version = "0.4.38", features = ["clock"], optional = true }
+chrono = { version = "0.4", features = ["clock"], optional = true }
+reqwest = { version = "0.12", features = ["blocking"] }
 
 [dev-dependencies]
 multi_log = "0.1.2"
+mockito = "1.6.1"
+serial_test = "3.2.0"
 
 [[example]]
 name = "log_to_discord"
